# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FarmaSystem - API de Gestão de Farmácia
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: Listar produtos
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: controlled
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Criar produto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Obter produto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Atualizar produto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Excluir produto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Excluído
  /products/barcode/{barcode}:
    get:
      operationId: getProductByBarcode
      tags: [products]
      summary: Buscar produto por código de barras
      parameters:
        - name: barcode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  # ─── Stock ────────────────────────────────────────────────────
...
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: Listar vendas
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Criar venda
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaleBody"
      responses:
        "201":
          description: Criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      summary: Obter venda
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}/cancel:
    patch:
      operationId: cancelSale
      tags: [sales]
      summary: Cancelar venda
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"

  # ─── Customers ────────────────────────────────────────────────
...
  /purchases:
    get:
      operationId: listPurchases
      tags: [purchases]
      summary: Listar pedidos de compra
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Purchase"
    post:
      operationId: createPurchase
      tags: [purchases]
      summary: Criar pedido de compra
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePurchaseBody"
      responses:
        "201":
          description: Criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
  /purchases/{id}:
    get:
      operationId: getPurchase
      tags: [purchases]
      summary: Obter pedido de compra
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
    patch:
      operationId: updatePurchase
      tags: [purchases]
      summary: Atualizar status do pedido
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePurchaseBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"

  # ─── Financial ────────────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        genericName:
          type: ["string", "null"]
        barcode:
          type: ["string", "null"]
        category:
          type: string
        manufacturer:
          type: ["string", "null"]
        dosage:
          type: ["string", "null"]
        unit:
          type: string
        costPrice:
          type: number
        salePrice:
          type: number
        minStock:
          type: integer
        currentStock:
          type: integer
        isControlled:
          type: boolean
        anvisaCode:
          type: ["string", "null"]
        requiresPrescription:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - category
        - unit
        - costPrice
        - salePrice
        - minStock
        - currentStock
        - isControlled
        - requiresPrescription
        - createdAt
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        genericName:
          type: ["string", "null"]
        barcode:
          type: ["string", "null"]
        category:
          type: string
        manufacturer:
          type: ["string", "null"]
        dosage:
          type: ["string", "null"]
        unit:
          type: string
        costPrice:
          type: number
        salePrice:
          type: number
        minStock:
          type: integer
        isControlled:
          type: boolean
        anvisaCode:
          type: ["string", "null"]
        requiresPrescription:
          type: boolean
      required:
        - name
        - category
        - unit
        - costPrice
        - salePrice
        - minStock
        - isControlled
        - requiresPrescription
...
    Sale:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        customerCpf:
          type: ["string", "null"]
        userId:
          type: integer
        userName:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        subtotal:
          type: number
        discount:
          type: number
        total:
          type: number
        paymentMethod:
          type: string
          enum: [dinheiro, cartao_credito, cartao_debito, pix, convenio]
        status:
          type: string
          enum: [concluida, cancelada, pendente]
        createdAt:
          type: string
      required:
        - id
        - userId
        - userName
        - items
        - subtotal
        - discount
        - total
        - paymentMethod
        - status
        - createdAt
    SaleItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        discount:
          type: number
        total:
          type: number
      required:
        - id
        - productId
        - productName
        - quantity
        - unitPrice
        - discount
        - total
    CreateSaleBody:
      type: object
      properties:
        customerId:
          type: ["integer", "null"]
        customerCpf:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateSaleItemBody"
        discount:
          type: number
        paymentMethod:
          type: string
          enum: [dinheiro, cartao_credito, cartao_debito, pix, convenio]
      required:
        - items
        - paymentMethod
    CreateSaleItemBody:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        discount:
          type: number
      required:
        - productId
        - quantity
...
    Purchase:
      type: object
      properties:
        id:
          type: integer
        supplierId:
          type: integer
        supplierName:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseItem"
        total:
          type: number
        status:
          type: string
          enum: [pendente, confirmado, recebido, cancelado]
        expectedDate:
          type: ["string", "null"]
        receivedDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - supplierId
        - supplierName
        - items
        - total
        - status
        - createdAt
    PurchaseItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitCost:
          type: number
        total:
          type: number
      required:
        - productId
        - productName
        - quantity
        - unitCost
        - total
    CreatePurchaseBody:
      type: object
      properties:
        supplierId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseItem"
        expectedDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - supplierId
        - items
    UpdatePurchaseBody:
      type: object
      properties:
        status:
          type: string
          enum: [pendente, confirmado, recebido, cancelado]
        receivedDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...